+++ html/index.html
@@ -165,7 +165,7 @@
     <script>
       require('../renderer.js');
       require('../js/nav.js');      // Navigation handling.
-      require('../js/home-test.js');     // Handle home page.
+      require('../js/home.js');     // Handle home page.
       require('../js/ticket-handler-repair.js');  // Handle network tickets / requests.
       require('../js/ticket-handler-errors.js');  // Handle form input errors.
     </script>
