+++ html/key.html
@@ -134,7 +134,7 @@
 
         <script>
             require('../js/nav.js');                    // Navigation handling.
-            require('../js/home-test.js');              // Handle home page.
+            require('../js/home.js');              // Handle home page.
             require('../js/ticket-handler-key.js');     // Handle network tickets / requests.
             require('../js/ticket-handler-errors.js');  // Handle form input errors.
         </script>
