+++ html/mtl.html
@@ -137,7 +137,7 @@
 
         <script>
             require('../js/nav.js');                    // Navigation handling.
-            require('../js/home-test.js');                   // Handle home page.
+            require('../js/home.js');                   // Handle home page.
             require('../js/ticket-handler-mtl.js');     // Handle network tickets / requests.
             require('../js/ticket-handler-errors.js');  // Handle form input errors.
         </script>
