+++ html/survey.html
@@ -142,7 +142,7 @@
 
     <script>
       require('../js/nav.js');      // Navigation handling.
-      require('../js/home-test.js');     // Handle home page.
+      require('../js/home.js');     // Handle home page.
       require('../js/ticket-handler-survey.js');  // Handle network tickets / requests.
       require('../js/ticket-handler-errors.js');  // Handle form input errors.
     </script>
