+++ html/voip.html
@@ -150,7 +150,7 @@
 
         <script>
             require('../js/nav.js');                    // Navigation handling
-            require('../js/home-test.js');                   // Handle home page
+            require('../js/home.js');                   // Handle home page
             require('../js/ticket-handler-voip.js');    // Handle network tickets / requests
             require('../js/ticket-handler-errors.js');  // Handle form input errors
         </script>
