+++ js/ticket-handler-onsite.js
@@ -1,4 +1,4 @@
-const Processes = require('./modules/Processes-test.js'); // All common processes with the form.
+const Processes = require('./modules/Processes.js'); // All common processes with the form.
 let Tickets = new function() // All form specific processes.
 {
     this.Handler = new function()
